Move rendering code out of documents models file
